requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
distlib==0.3.8
Django==5.0.6
django-bootstrap3==24.2
django-debug-toolbar==4.3.0
django-filter==24.2
django-import-export==4.0.3
django-pandas==0.6.7
django-redis==5.4.0
django-tables2==2.7.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
filelock==3.14.0
gunicorn==22.0.0
hiredis==2.3.2
mysql-connector-python==8.4.0
mysqlclient==2.2.4
npm==0.1.1
numpy==1.26.4
openpyxl==3.1.2
optional-django==0.1.0
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
virtualenv==20.26.2
